@@ -1,19 +1,30 @@
 ﻿using System;
+using System.ComponentModel;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Serilog;
+using PropertyChanged;
 
 namespace SWTools.Core {
     /// <summary>
     /// 创意工坊物品
     /// </summary>
-    public partial class Item {
-        public string ItemId { get; set; } = "";      // 物品 Id
-        public string ItemTitle { get; set; } = "";   // 物品标题
-        public long ItemSize { get; set; }            // 该物品的文件大小
-        public long AppId { get; set; }               // 物品属于的 App 的 Id
-        public string AppName { get; set; } = "";     // 物品属于的 App 的名字
+    [AddINotifyPropertyChangedInterface]
+    public partial class Item : INotifyPropertyChanged {
+        public event PropertyChangedEventHandler? PropertyChanged;
+        protected virtual void OnPropertyChanged(string propertyName) {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+        public string ItemId { get; set; } = string.Empty;      // 物品 Id
+        public string ItemTitle { get; set; } = string.Empty;   // 物品标题
+        public long ItemSize { get; set; }                      // 该物品的文件大小
+        public long AppId { get; set; }                         // 物品属于的 App 的 Id
+        public string AppName { get; set; } = string.Empty;     // 物品属于的 App 的名字
+        public string UrlPreview { get; set; } = string.Empty;  // 预览文件地址
+        // 有些物品不需要 Steamcmd 即可下载
+        public bool IsFree { get; set; }                              // 是否不需要 Steamcmd
+        public string UrlFreeDownload { get; set; } = string.Empty;   // 下载地址
+
 
         // 解析状态
         [JsonConverter(typeof(JsonStringEnumConverter))]
@@ -26,11 +37,11 @@ namespace SWTools.Core {
         // 下载失败原因
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public EFailReason FailReason { get; set; } = EFailReason.Null;
-        private string _exceptionMsg = "";
+        private string _exceptionMsg = string.Empty;
 
         // 解析之后的操作
         [JsonIgnore]
-        public EAfterParse AfterParse { get; set; } = EAfterParse.Nothing; 
+        public EAfterParse AfterParse { get; set; } = EAfterParse.Nothing;
 
         public enum EParseState {
             InQueue, Handling, Done, Failed,
@@ -42,14 +53,12 @@ namespace SWTools.Core {
         public enum EFailReason {
             Null, Unknown, Exception,
             FileNotFound, Timeout, NoConnection,
-            AccountDisabled, InvalidPassword, NoMatch
+            AccountDisabled, InvalidPassword, NoMatch,
+            AccessDenied, LockingFailed
         }
         public enum EAfterParse {
             Nothing, Download
         }
-
-        // 空物品
-        public static readonly Item Empty = new();
 
         public Item() { }
         public Item(string itemId) {
@@ -59,10 +68,10 @@ namespace SWTools.Core {
         // 序列化到 Json
         public override string ToString() {
             try {
-                return JsonSerializer.Serialize(this, Helper._jsonOptions);
+                return JsonSerializer.Serialize(this, Constants.JsonOptions);
             }
             catch (Exception ex) {
-                Log.Logger.Error("Exception occured when serializing Json:\n{Exception}", ex);
+                LogManager.Log.Error("Exception occured when serializing Json:\n{Exception}", ex);
                 return string.Empty;
             }
         }
@@ -75,7 +84,7 @@ namespace SWTools.Core {
                 case EFailReason.Exception:
                     return "异常：" + _exceptionMsg;
                 case EFailReason.Unknown:
-                    return "未知错误";
+                    return "未知错误（查看日志获取更多信息）";
                 case EFailReason.FileNotFound:
                     return "物品未找到，请检查物品 ID 是否正确";
                 case EFailReason.Timeout:
@@ -88,18 +97,18 @@ namespace SWTools.Core {
                     return "密码错误，请尝试向开发者反映此问题";
                 case EFailReason.NoMatch:
                     return "Steam App 与物品 ID 不匹配";
+                case EFailReason.AccessDenied:
+                    return "账户权限不足";
+                case EFailReason.LockingFailed:
+                    return "文件被占用";
             }
-            Log.Logger.Error("Received unknown enum value: {FailReason}", FailReason);
+            LogManager.Log.Error("Received unknown enum value: {FailReason}", FailReason);
             return string.Empty;
         }
 
         // 获取下载文件目录
         public string GetDownloadPath() {
-            StringBuilder sb = new();
-            sb.Append("steamcmd\\steamapps\\workshop\\content\\");
-            sb.Append(AppId).Append('\\');
-            sb.Append(ItemId).Append('\\');
-            return sb.ToString();
+            return Constants.SteamcmdDir + $"steamapps/workshop/content/{AppId}/{ItemId}/";
         }
 
         // 把状态二值化到 InQueue / Done
@@ -126,8 +135,32 @@ namespace SWTools.Core {
                 return EFailReason.InvalidPassword;
             } else if (downloadLog.Contains("No match")) {
                 return EFailReason.NoMatch;
+            } else if (downloadLog.Contains("Access Denied")) {
+                return EFailReason.AccessDenied;
+            } else if (downloadLog.Contains("Locking Failed")) {
+                return EFailReason.LockingFailed;
             }
             return EFailReason.Unknown;
+        }
+
+        // 解析自己 (注意: 如果要解析多个物品，请使用 ItemList.ParseAll(), 此方法主要用于测试)
+        public async Task Parse() {
+            ParseState = EParseState.Handling;
+            var response = await API.SwDownloader.Request([ItemId]);
+            if (response == null || response?.Length == 0) {
+                LogManager.Log.Error("Failed to parse item {ItemId}", ItemId);
+                return;
+            }
+            ParseWith(response![0]);
+        }
+
+        // 检查信息是否完备
+        public bool IsCompleted() {
+            return !string.IsNullOrEmpty(ItemId) &&
+                !string.IsNullOrEmpty(ItemTitle) &&
+                ItemSize != 0 && AppId != 0 &&
+                !string.IsNullOrEmpty(AppName) &&
+                !string.IsNullOrEmpty(UrlPreview);
         }
     }
 }
